requirements.txt
Django==1.8
django-filter==0.8
djangorestframework==3.0.5
psycopg2==2.5.4
pytz==2016.10
six==1.10.0
sqlparse==0.2.2
wheel==0.24.0
